add Right to edit and delete

+++ controller/posts/delete.js
@@ -1,4 +1,5 @@
 const { post } = require('../../models');
+const { user } = require('../../models');
 const jwt = require('jsonwebtoken');
 
 module.exports = {
@@ -14,19 +15,42 @@ module.exports = {
     //로그인 세션이 연결되어있다면
     if (email) {
       post
-        .destroy({
-          //post를 삭제하므로 cascade로 연결되어있는 다른 테이블의 값들도 자동 삭제
+        .findOne({
           where: {
-            id: id, //전달받은 posts의 id를 검사
+            id: id,
           },
         })
-        .then((result) => {
-          //존재할 시 삭제
-          if (result) {
-            res.status(200).send('delete post');
-          } else {
-            res.status(400).send('Invalid Request');
-          }
+        .then((findPost) => {
+          user
+            .findOne({
+              where: {
+                id: findPost.dataValues.authorId,
+              },
+            })
+            .then((findUser) => {
+              if (findUser.dataValues.email === email) {
+                post
+                  .destroy({
+                    //post를 삭제하므로 cascade로 연결되어있는 다른 테이블의 값들도 자동 삭제
+                    where: {
+                      id: id, //전달받은 posts의 id를 검사
+                    },
+                  })
+                  .then((result) => {
+                    //존재할 시 삭제
+                    if (result) {
+                      res.status(200).send('delete post');
+                    } else {
+                      res.status(400).send('Invalid Request');
+                    }
+                  })
+                  .catch(() => {
+                    res.sendStatus(500);
+                  });
+              } else {
+                res.status(400).send('Invalid Request');
+              }
+            });
         })
         .catch(() => {
           res.sendStatus(500);
